Refactor body-parser
