+++ lib/error-handler.js
@@ -1,8 +1,8 @@
 'use strict';
 
 module.exports = function errorHandler(err, req, res, next) {
-    const code = error.code || 500;
-    const error = code === 500 ? 'Internal Server Error' : error.error;
+    const code = err.code || 500;
+    const error = code === 500 ? 'Internal Server Error' : err.error;
     console.error(err.error || err.message);
         res.status(code).send({ error });
 };
