+++ lib/models/breeds.js
@@ -1,5 +1,5 @@
 'use strict';
 
-const Breed = require('./storageScoutFiles');
+const Breed = require('./memory-store');
 
 module.exports = new Breed();
