+++ lib/models/memory-store.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const storageScout = null;
+
 module.exports = class Breed {
     constructor() {
         this.data = [];
@@ -37,7 +39,7 @@ module.exports = class Breed {
     delete(id) {
         return new Promise((resolve, reject) => {
             const index = this.data.findIndex(u => u.id === id);
-            if (index! == -1) {
+            if (index !== -1) {
                 const deleted = this.data.splice(index, 1)[0];
                 resolve(deleted);
             }
